+++ CHANGELOG.rst
@@ -7,6 +7,13 @@ All notable changes to this project will be documented in this file.
 The format is based on `Keep a Changelog <http://keepachangelog.com/en/1.0.0/>`_
 and this project adheres to `Semantic Versioning <http://semver.org/spec/v2.0.0.html>`_.
 
+Unreleased
+----------
+
+**Changed**
+
+* For consistency, ``timeout`` parameters of ``Goal.send()`` and ``ActionClient()`` are now expressed in **seconds**, instead of milliseconds.
+
 0.5.0
 ----------